report a timestamp count the SDKs cannot represent
The four timestamp constructors take a signed count and report one outside
the microsecond range a timestamp covers: C returns a GtdStatus and writes a
GtdTimestamp out parameter, Rust returns a Result with the new
Error::TimestampCountOutOfRange, and C++ has a try_* and a throwing pair.
The nanosecond constructors truncate towards zero to whole microseconds.

+++ sdk/cpp/examples/from_multiple_sources.cpp
@@ -22,11 +22,11 @@
 namespace {
 // 2024-06-01T08:00:00Z. File-scope so the `at` lambda can read it without a
 // capture (MSVC C++17 requires capturing a local).
-constexpr std::uint64_t kBase = 1717228800;
+constexpr std::int64_t kBase = 1717228800;
 } // namespace
 
 int main() {
-    auto timestamp_at = [](std::uint64_t secs) {
+    auto timestamp_at = [](std::int64_t secs) {
         return geotrace::Timestamp::from_seconds(kBase + secs);
     };
 
@@ -48,7 +48,7 @@ int main() {
             {51.5078, -0.1270, 90.0},
             {51.5079, -0.1268, 90.5},
         }};
-        std::size_t idx = 0;
+        std::int64_t idx = 0;
         for (const auto &point : gps) {
             geotrace::NavFix fix{geotrace::FixTime::receiver(timestamp_at(idx * 10)),
                                  geotrace::Angle::degrees(point.lat),
@@ -61,7 +61,7 @@ int main() {
         // Source 2: annotations from a separate log.  Their map positions are
         // not supplied - finish() interpolates them from the GPS fixes by time.
         struct Marker {
-            std::uint64_t offset;
+            std::int64_t offset;
             std::string_view label;
             geotrace::MarkerIcon icon;
         };
